requirements.txt
cryptography
pycryptodome
matplotlib
seaborn
pandas
numpy
psutil
tabulate
scipy
